requirements.txt
numpy
pillow
h5py
tqdm